feat: Materials are "working"!
This commit makes a single mesh with a single material work.
We have a long way to go to make this work in general, but a start is
a start!

We have disabled any interaction with the previous system, i.e. the old
descriptor sets, since they throw validation errors that crash the
application.